# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Carga Horaria - API",
        "description": "Documentación completa de la API del sistema de gestión de carga horaria docente",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/teachers": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU06 - Listar docentes",
                "description": "Lista todos los docentes registrados en el sistema",
                "operationId": "9115585a972eee31c1bb78305dc1620c",
                "responses": {
                    "200": {
                        "description": "Lista de docentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email"
                                            },
                                            "dni": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "phone": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "department": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Solo administradores"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU06 - Crear docente",
                "description": "Registra un nuevo docente en el sistema",
                "operationId": "232f4cec049f7f25ccdce549242471bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@universidad.edu"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+591 70123456",
                                        "nullable": true
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "Ingeniería de Sistemas",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Docente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validación fallida - Email duplicado"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/teachers/{id}": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU06 - Ver docente",
                "description": "Obtiene los detalles de un docente específico",
                "operationId": "99560f14c85bbc9bc2411477de3479dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "dni": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "department": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU06 - Eliminar docente",
                "description": "Elimina un docente del sistema (solo administradores)",
                "operationId": "8cdf35012d9c8278e43bb384ce1d70db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docente eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Teacher deleted"
                                        },
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU06 - Actualizar docente (Admin)",
                "description": "Actualiza los datos de un docente (solo administradores)",
                "operationId": "b1e4694132990a3485ac634c072a8a47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "department": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Docente actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Docente no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/api/teachers/me": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU07 - Ver mi perfil de docente",
                "description": "El docente autenticado visualiza su propio perfil",
                "operationId": "d0df8488ceb40267a13fcf25fb02a385",
                "responses": {
                    "200": {
                        "description": "Perfil del docente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "dni": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "department": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Perfil de docente no encontrado"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Docentes"
                ],
                "summary": "CU07 - Actualizar mi perfil de docente",
                "description": "El docente autenticado edita su propio perfil",
                "operationId": "4fc787e926b1b7f2a0549d66b036b57d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "department": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}